Move config/dirs to helpers to avoid any confusion with application configuration
