+++ framework/src/controller/controller.js
@@ -2,7 +2,7 @@ const fs = require('fs-extra');
 const path = require('path');
 const child_process = require('child_process');
 const psList = require('ps-list');
-const systemDirs = require('./config/dirs');
+const systemDirs = require('./helpers/sysmtem_dirs');
 const { InMemoryChannel } = require('./channels');
 const Bus = require('./bus');
 const { DuplicateAppInstanceError } = require('../errors');
